refactor: remove mock/registry dir
