+++ test/collections/componentSet.test.ts
@@ -25,9 +25,9 @@ import {
 import { ComponentSetError } from '../../src/errors';
 import { nls } from '../../src/i18n';
 import { mockConnection } from '../mock/client';
-import { decomposedtoplevel, matchingContentFile, mixedContentSingleFile } from '../mock/registry';
-import { MATCHING_RULES_COMPONENT } from '../mock/registry/type-constants/nonDecomposedConstants';
-import * as manifestFiles from '../mock/registry/manifestConstants';
+import { decomposedtoplevel, matchingContentFile, mixedContentSingleFile } from '../mock';
+import { MATCHING_RULES_COMPONENT } from '../mock/type-constants/nonDecomposedConstants';
+import * as manifestFiles from '../mock/manifestConstants';
 
 const env = createSandbox();
 const $$ = testSetup(env);
