+++ test/convert/transformers/decomposedMetadataTransformer.ts
@@ -9,11 +9,11 @@ import { join } from 'path';
 import { expect } from 'chai';
 import { createSandbox } from 'sinon';
 import { assert } from '@salesforce/ts-types';
-import { decomposed, matchingContentFile } from '../../mock/registry';
+import { decomposed, matchingContentFile } from '../../mock';
 import { DecomposedMetadataTransformer } from '../../../src/convert/transformers/decomposedMetadataTransformer';
 import { baseName } from '../../../src/utils';
 import { JsToXml } from '../../../src/convert/streams';
-import { DECOMPOSED_TOP_LEVEL_COMPONENT } from '../../mock/registry/type-constants/decomposedTopLevelConstants';
+import { DECOMPOSED_TOP_LEVEL_COMPONENT } from '../../mock/type-constants/decomposedTopLevelConstants';
 import { ComponentSet, ForceIgnore, registry, RegistryAccess, SourceComponent } from '../../../src';
 import { XML_NS_KEY, XML_NS_URL } from '../../../src/common';
 import { ConvertContext } from '../../../src/convert/convertContext';
