+++ test/mock/type-constants/index.ts
