+++ test/mock/type-constants/matchingContentFileConstants.ts
@@ -5,8 +5,8 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { join } from 'path';
-import { registry, SourceComponent, VirtualTreeContainer } from '../../../../src';
-import { META_XML_SUFFIX } from '../../../../src/common';
+import { registry, SourceComponent, VirtualTreeContainer } from '../../../src';
+import { META_XML_SUFFIX } from '../../../src/common';
 
 // Constants for a matching content file type (ApexClass)
 const type = registry.types.apexclass;
