+++ test/mock/type-constants/mixedContentDirectoryConstants.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { basename, dirname, join } from 'path';
-import { registry, SourceComponent } from '../../../../src';
+import { registry, SourceComponent } from '../../../src';
 
 // Mixed content with directory as content
 const type = registry.types.staticresource;
