+++ test/mock/type-constants/mixedContentSingleFileConstants.ts
@@ -6,8 +6,8 @@
  */
 import { join } from 'path';
 
-import { registry, SourceComponent } from '../../../../src';
-import { META_XML_SUFFIX } from '../../../../src/common';
+import { registry, SourceComponent } from '../../../src';
+import { META_XML_SUFFIX } from '../../../src/common';
 
 const type = registry.types.staticresource;
